test: separate unit test and inegration test
