+++ promptulate/utils/core_utils.py
@@ -108,37 +108,35 @@ def get_cache():
     return cache
 
 
-def set_openai_api_key(value: str):
-    cache["OPENAI_API_KEY"] = value
-
-
 def convert_backslashes(path: str):
     """Convert all \\ to / of file path."""
     return path.replace("\\", "/")
 
 
 def get_default_storage_path(module_name: str = "") -> str:
-    """Get the default storage path for the current module. The storage path is
-    created in the user's home directory, or in a temporary directory if permission
-    is denied.
+    """Get the default storage path for the current module.
+
+    The storage path is created in the user's home directory under ~/.zeeland/pne,
+    or in a temporary directory if permission is denied.
 
     Args:
-        module_name(str): The name of the module to create a storage path for.
+        module_name (str): The name of the module to create a storage path for.
 
     Returns:
         str: The default storage path for the current module.
     """
-    storage_path = os.path.expanduser("~/.pne")
+    storage_path = os.path.expanduser("~/.zeeland/pne")
 
     if module_name:
         storage_path = os.path.join(storage_path, module_name)
 
-    # Try to create the storage path (with module subdirectory if specified)
-    # Use a temporary directory instead if permission is denied,
     try:
         os.makedirs(storage_path, exist_ok=True)
     except PermissionError:
-        storage_path = os.path.join(tempfile.gettempdir(), "pne", module_name)
+        temp_path = os.path.join(tempfile.gettempdir(), "zeeland", "pne")
+        storage_path = (
+            os.path.join(temp_path, module_name) if module_name else temp_path
+        )
         os.makedirs(storage_path, exist_ok=True)
 
     return convert_backslashes(storage_path)
