+++ tests/old/output_formatter/test_formatter.py
