+++ tests/old/tools/test_Calculator_class.py
